Ajout des clés étrangères de la table `sport_articles` aux migrations.

+++ database/migrations/2019_03_13_122354_create_sport_articles_table.php
@@ -15,10 +15,10 @@ class CreateSportArticlesTable extends Migration
     {
         Schema::create('sport_articles', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('sport_id');
-            $table->unsignedInteger('club_id');
-            $table->unsignedInteger('article_id');
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('sport_id');
+            $table->unsignedBigInteger('club_id');
+            $table->unsignedBigInteger('article_id');
+            $table->unsignedBigInteger('user_id');
         });
     }
